+++ app/src/api/v1/payment_methods.py
@@ -1,3 +1,6 @@
+from http import HTTPStatus
+from typing import Annotated
+
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi_pagination import Page, paginate
 from schemas.payment_method import (CreatePaymentMethodSchema,
@@ -8,6 +11,7 @@ from services.exceptions import (ConflictError, ObjectAlreadyExistsException,
 from services.payment_method_service import (PaymentMethodService,
                                              get_payment_method_service)
 from sqlalchemy.exc import DBAPIError
+from utils.auth import check_user_access, decode_token, oauth2_scheme
 
 router = APIRouter()
 
@@ -15,9 +19,19 @@ router = APIRouter()
 @router.post("/", response_model=GetPaymentMethodSchema)
 async def create_payment_method(
     payment_method: CreatePaymentMethodSchema,
+    access_token: Annotated[str, Depends(oauth2_scheme)],
     payment_method_service: PaymentMethodService = Depends(get_payment_method_service),
 ) -> GetPaymentMethodSchema:
     try:
+        payload = decode_token(access_token)
+
+        if not payload:
+            raise HTTPException(
+                status_code=HTTPStatus.UNAUTHORIZED, detail="invalid token"
+            )
+
+        check_user_access(payload, str(payment_method.user_id))
+
         payment_method = await payment_method_service.create_payment_method(
             payment_method
         )
@@ -30,12 +44,22 @@ async def create_payment_method(
 
 
 @router.patch("/{payment_method_id}", response_model=GetPaymentMethodSchema)
-async def update_category(
+async def update_payment_method(
     payment_method_id: str,
+    access_token: Annotated[str, Depends(oauth2_scheme)],
     payment_method: UpdatePaymentMethodSchema,
     payment_method_service: PaymentMethodService = Depends(get_payment_method_service),
 ) -> GetPaymentMethodSchema:
     try:
+        payload = decode_token(access_token)
+
+        if not payload:
+            raise HTTPException(
+                status_code=HTTPStatus.UNAUTHORIZED, detail="invalid token"
+            )
+
+        check_user_access(payload, str(payment_method.user_id))
+
         updated_payment_method = await payment_method_service.update_payment_method(
             payment_method_id, payment_method
         )
@@ -50,9 +74,19 @@ async def update_category(
 @router.get("/{payment_method_id}", response_model=GetPaymentMethodSchema)
 async def get_payment_method_by_id(
     payment_method_id: str,
+    access_token: Annotated[str, Depends(oauth2_scheme)],
     payment_method_service: PaymentMethodService = Depends(get_payment_method_service),
 ) -> GetPaymentMethodSchema:
     try:
+        payload = decode_token(access_token)
+
+        if not payload:
+            raise HTTPException(
+                status_code=HTTPStatus.UNAUTHORIZED, detail="invalid token"
+            )
+
+        check_user_access(payload, str(payload["user_id"]))
+
         payment_method = await payment_method_service.get_payment_method_by_id(
             payment_method_id
         )
@@ -67,9 +101,19 @@ async def get_payment_method_by_id(
 @router.get("/users/{user_id}", response_model=Page[GetPaymentMethodSchema])
 async def get_payment_method_by_user_id(
     user_id: str,
+    access_token: Annotated[str, Depends(oauth2_scheme)],
     payment_method_service: PaymentMethodService = Depends(get_payment_method_service),
 ) -> Page[GetPaymentMethodSchema]:
     try:
+        payload = decode_token(access_token)
+
+        if not payload:
+            raise HTTPException(
+                status_code=HTTPStatus.UNAUTHORIZED, detail="invalid token"
+            )
+
+        check_user_access(payload, user_id)
+
         payment_methods = await payment_method_service.get_payment_methods_by_user_id(
             user_id
         )
@@ -87,9 +131,19 @@ async def get_payment_method_by_user_id(
 )
 async def delete_payment_method(
     payment_method_id: str,
+    access_token: Annotated[str, Depends(oauth2_scheme)],
     payment_method_service: PaymentMethodService = Depends(get_payment_method_service),
 ):
     try:
+        payload = decode_token(access_token)
+
+        if not payload:
+            raise HTTPException(
+                status_code=HTTPStatus.UNAUTHORIZED, detail="invalid token"
+            )
+
+        check_user_access(payload, str(payload["user_id"]))
+
         await payment_method_service.delete_payment_method(payment_method_id)
         return {"detail": "success"}
     except ObjectNotFoundError as error:
